refactor: renames "index.ts" to "definition.ts" in "features/TextToImage/types/Output"
